+++ Components/MeshLodGenerator/src/OgreLodOutputProviderCompressedMesh.cpp
@@ -271,7 +271,7 @@ namespace Ogre
         }
 
         // Close buffers.
-        for (unsigned short i = 0; i < submeshCount; i++)
+        for (unsigned i = 0; i < submeshCount; i++)
         {
             v1::SubMesh::LODFaceList& lods = mMesh->getSubMesh(i)->mLodFaceList[VpNormal];
             v1::IndexData* prevLod = lods[mLastIndexBufferID];
