+++ Components/MeshLodGenerator/src/OgreLodOutputProviderMesh.cpp
@@ -139,7 +139,7 @@ namespace Ogre
         }
 
         // Close buffers.
-        for (unsigned short i = 0; i < submeshCount; i++)
+        for (unsigned i = 0; i < submeshCount; i++)
         {
             v1::SubMesh::LODFaceList& lods = mMesh->getSubMesh(i)->mLodFaceList[VpNormal];
             lods[lodIndex]->indexBuffer->unlock();
